update alerts to sweet alerts in course prerequisites.

+++ scripts/course_prerequisites.js
@@ -57,7 +57,7 @@ function addCoursePrerequisite() {
   })
   .then((response) => response.text())
   .then(responseText => {
-    alert(responseText);
+    Swal.fire("Success", `${responseText}`, "success");
     displayPrerequisiteCourses(); // updates the table
   }).catch (error => {
     alert('console error.');
@@ -98,7 +98,7 @@ function editCoursePrerequisite(button) {
     })
     .then((response) => response.text())
     .then((responseText) => {
-      alert(responseText);
+      Swal.fire("Success", `${responseText}`, "success");
       displayPrerequisiteCourses();
     });
 
@@ -107,6 +107,28 @@ function editCoursePrerequisite(button) {
     // Edit mode
     button.textContent = "Save";
     cells[0].focus();
+
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      icon: "info",
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      customClass: {
+        popup: 'colored-toast'
+      },
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      }
+    });
+
+    Toast.fire({
+      icon: 'info',
+      title: 'Edit Mode Activated'
+    });
+
     updatedRow.length = 0; // clear the array.
   }
 }
@@ -116,19 +138,28 @@ function deleteCoursePrerequisite(button) {
   const row = button.closest('tr');
   const cells = row.querySelectorAll('td');
   const courseId = cells[0].innerHTML;
-  
-  fetch(coursePrerequisiteEndpoint, {
-    method: 'DELETE',
-    headers: {
-      "Content-type": "application/x-www-form-urlencoded",
-    },
-    body: `course_id=${courseId}`,
-  })
-  .then((response) => response.text())
-  .then((responseText) => {
-    alert(responseText);
-    displayPrerequisiteCourses();
-  })
+
+  Swal.fire({
+    title: "Do you want to delete this data?",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonText: "Delete",
+  }).then((result) => {
+    if (result.isConfirmed) {
+        fetch(coursePrerequisiteEndpoint, {
+          method: 'DELETE',
+          headers: {
+            "Content-type": "application/x-www-form-urlencoded",
+          },
+          body: `course_id=${courseId}`,
+        })
+        .then((response) => response.text())
+        .then(() => {
+          Swal.fire("Deleted", "", "success");
+          displayPrerequisiteCourses();
+      })
+    }
+  });
 }
 
 function searchCourse() {
